client/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@0551214be1957a307d0bd3a9ef5056f0fd9e58bc
frozenlist==1.3.0
idna==3.3
loguru==0.6.0
multidict==6.0.2
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pyzmq==23.0.0
six==1.16.0
typing_extensions==4.2.0
yarl==1.7.2
